Moved the FluorophoreDynamics-related classes to a new photophysics package.
