+++ src/main/java/ch/epfl/leb/sass/models/fluorophores/internal/commands/GenerateFluorophoresGrid3D.java
@@ -18,7 +18,7 @@
 package ch.epfl.leb.sass.models.fluorophores.internal.commands;
 
 import ch.epfl.leb.sass.models.fluorophores.Fluorophore;
-import ch.epfl.leb.sass.models.fluorophores.internal.dynamics.FluorophoreDynamics;
+import ch.epfl.leb.sass.models.photophysics.FluorophoreDynamics;
 import ch.epfl.leb.sass.models.components.Camera;
 import ch.epfl.leb.sass.models.psfs.PSFBuilder;
 import java.util.List;
