+++ src/main/java/ch/epfl/leb/sass/models/legacy/dStormProperties.java
@@ -21,7 +21,7 @@ package ch.epfl.leb.sass.models.legacy;
 
 import ch.epfl.leb.sass.models.legacy.Camera;
 import ch.epfl.leb.sass.models.fluorophores.internal.DefaultFluorophore;
-import ch.epfl.leb.sass.models.fluorophores.internal.StateSystem;
+import ch.epfl.leb.sass.models.photophysics.StateSystem;
 import ch.epfl.leb.sass.models.legacy.Fluorophore3D;
 import ch.epfl.leb.sass.models.legacy.FluorophoreProperties;
 import ch.epfl.leb.sass.models.psfs.PSFBuilder;
