+++ src/main/java/ch/epfl/leb/sass/models/photophysics/FluorophoreDynamics.java
@@ -15,9 +15,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.sass.models.fluorophores.internal.dynamics;
+package ch.epfl.leb.sass.models.photophysics;
 
-import ch.epfl.leb.sass.models.fluorophores.internal.StateSystem;
 import java.io.Serializable;
 
 /**
