+++ src/main/java/ch/epfl/leb/sass/models/photophysics/StateSystem.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package ch.epfl.leb.sass.models.fluorophores.internal;
+package ch.epfl.leb.sass.models.photophysics;
 
 import java.io.Serializable;
