+++ src/test/java/ch/epfl/leb/sass/models/MicroscopeIT.java
@@ -20,7 +20,7 @@ package ch.epfl.leb.sass.models;
 import ch.epfl.leb.sass.IntegrationTest;
 import ch.epfl.leb.sass.models.components.*;
 import ch.epfl.leb.sass.models.psfs.internal.Gaussian2D;
-import ch.epfl.leb.sass.models.fluorophores.internal.dynamics.PalmDynamics;
+import ch.epfl.leb.sass.models.photophysics.internal.PalmDynamics;
 import ch.epfl.leb.sass.models.fluorophores.internal.commands
                                                     .GenerateFluorophoresGrid2D;
 import ch.epfl.leb.sass.models.obstructors.internal.commands
